feat(news-app): create pages layout
