+++ packages/news-app/pages/index.tsx
@@ -1,20 +1,8 @@
-import { useMediaQuery } from "@mui/material";
+import Layout from "@components/Layout";
 import type { NextPage } from "next";
-import { device } from "@assets/styles/breakpoints";
-import Header from "@components/Header";
-import Nav from "@components/Nav";
-import DrawerMenu from "@components/Nav/DrawerMenu";
-import { DrawerStorage } from "@context/DrawerContext";
 
 const Home: NextPage = () => {
-  const isTablet = useMediaQuery(device.tablet);
-  return (
-    <DrawerStorage>
-      <Header />
-      {!isTablet && <DrawerMenu />}
-      {isTablet && <Nav />}
-    </DrawerStorage>
-  );
+  return <Layout>Home</Layout>;
 };
 
 export default Home;
